Refactor: Split file wipe logic into modular static utility classes
