--- Services/DeleteBase.cs
@@ -1,207 +1,0 @@
-﻿using Spectre.Console;
-using System.Security.Cryptography;
-using ZeroTrace.Enums;
-using ZeroTrace.Model;
-
-namespace ZeroTrace.Services;
-
-public abstract class DeleteBase
-{
-    protected static int GetBufferSize(long fileLength = 0)
-    {
-        var defaultBufferSize = 64 * 1024 * 1024;
-
-        if (fileLength == 0)
-            return defaultBufferSize;
-
-        if (fileLength < 1 * 1024 * 1024)
-            return 64 * 1024;
-
-        if (fileLength < 10 * 1024 * 1024)
-            return 512 * 1024;
-
-        return defaultBufferSize;
-    }
-    protected static void Reverse(FileStream stream)
-    {
-        int left = 0;
-        long right = stream.Length - 1;
-
-        int bufferSize = GetBufferSize(stream.Length);
-        byte[] leftBuffer = new byte[bufferSize];
-        byte[] rightBuffer = new byte[bufferSize];
-
-        while (left < right)
-        {
-            int blockSize = (int)Math.Min(GetBufferSize(stream.Length), (right - left + 1) / 2);
-
-            stream.Position = left;
-            stream.ReadExactly(leftBuffer, 0, blockSize);
-
-            stream.Position = right - blockSize + 1;
-            stream.ReadExactly(rightBuffer, 0, blockSize);
-
-            stream.Position = left;
-            for (int i = blockSize - 1; i >= 0; i--)
-            {
-                stream.WriteByte(rightBuffer[i]);
-            }
-
-            stream.Position = right - blockSize + 1;
-            for (int i = blockSize - 1; i >= 0; i--)
-            {
-                stream.WriteByte(leftBuffer[i]);
-            }
-
-            left += blockSize;
-            right -= blockSize;
-        }
-
-        stream.Flush();
-    }
-
-    protected static void Shake(FileStream stream, RandomNumberGenerator rng, int intensity = 500)
-    {
-        long length = stream.Length;
-        if (length < 2) return;
-
-        for (int i = 0; i < intensity; i++)
-        {
-            long pos1 = GetRandomLong(rng, 0, length);
-            long pos2 = GetRandomLong(rng, 0, length);
-
-            stream.Position = pos1;
-            int b1 = stream.ReadByte();
-            stream.Position = pos2;
-            int b2 = stream.ReadByte();
-
-            if (b1 == -1 || b2 == -1) continue;
-
-            stream.Position = pos1;
-            stream.WriteByte((byte)b2);
-            stream.Position = pos2;
-            stream.WriteByte((byte)b1);
-        }
-
-        stream.Flush();
-    }
-
-    protected static void ZeroFill(FileStream stream)
-    {
-        var bufferSize = GetBufferSize(stream.Length);
-        byte[] buffer = new byte[bufferSize];
-        long length = stream.Length;
-        stream.Position = 0;
-
-        while (length > 0)
-        {
-            int toWrite = (int)Math.Min(bufferSize, length);
-            stream.Write(buffer, 0, toWrite);
-            length -= toWrite;
-        }
-
-        stream.Flush();
-    }
-
-    protected static void Random(FileStream stream, RandomNumberGenerator rng)
-    {
-        var bufferSize = GetBufferSize(stream.Length);
-        byte[] buffer = new byte[bufferSize];
-        long length = stream.Length;
-        stream.Position = 0;
-
-        while (length > 0)
-        {
-            int toWrite = (int)Math.Min(bufferSize, length);
-            rng.GetBytes(buffer, 0, toWrite);
-            stream.Write(buffer, 0, toWrite);
-            length -= toWrite;
-        }
-
-        stream.Flush();
-    }
-
-    protected static int GetRandomInt(RandomNumberGenerator rng, int min, int max)
-    {
-        ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(min, max);
-
-        byte[] bytes = new byte[4];
-        int range = max - min;
-        int result;
-
-        do
-        {
-            rng.GetBytes(bytes);
-            result = BitConverter.ToInt32(bytes, 0) & int.MaxValue;
-        }
-        while (result >= (int.MaxValue - (int.MaxValue % range)));
-
-        return min + (result % range);
-    }
-    protected static long GetRandomLong(RandomNumberGenerator rng, long min, long max)
-    {
-        ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(min, max);
-
-        byte[] bytes = new byte[8];
-        long range = max - min;
-        long result;
-
-        do
-        {
-            rng.GetBytes(bytes);
-            result = BitConverter.ToInt64(bytes, 0) & long.MaxValue;
-        } while (result >= (long.MaxValue - (long.MaxValue % range)));
-
-        return min + (result % range);
-    }
-
-    public static List<string> GetFiles(List<TargetItem> targets)
-    {
-        List<string> filePaths = [];
-
-        foreach (var target in targets)
-        {
-            if (target.Type != TargetType.File)
-            {
-                try
-                {
-                    filePaths.AddRange(GetSafeFiles(target.Path));
-                }
-                catch (Exception ex)
-                {
-                    AnsiConsole.MarkupLine($"[red]Error accessing {target.Path}: {ex.Message}[/]");
-                }
-            }
-            else
-            {
-                filePaths.Add(target.Path);
-            }
-        }
-
-        return filePaths;
-    }
-
-    private static List<string> GetSafeFiles(string path)
-    {
-        var files = new List<string>();
-        try
-        {
-            files.AddRange(Directory.GetFiles(path));
-            foreach (var dir in Directory.GetDirectories(path))
-            {
-                try
-                {
-                    files.AddRange(GetSafeFiles(dir));
-                }
-                catch (UnauthorizedAccessException)
-                {
-                }
-            }
-        }
-        catch (UnauthorizedAccessException)
-        {
-        }
-
-        return files;
-    }
-}
